add viewTrip route
