+++ src/App.jsx
@@ -7,6 +7,7 @@ import TripPage from './Components/TripPage/TripPage';
 import { SkeletonCircle } from '@chakra-ui/react';
 import { useSelector } from 'react-redux';
 import MyProfilePage from './Components/MyProfilePage/MyProfilePage';
+import ViewTripPage from './Components/TripPage/ViewTripPage';
 
 export default function App() {
   const lib = useSelector((state) => state.trip.mapsLibraries);
@@ -24,6 +25,7 @@ export default function App() {
       <Routes>
         <Route path="/" element={<HomePage />} />
         <Route path="/trip" element={<TripPage />} />
+        <Route path="/trip/:id" element={<ViewTripPage />} />
         <Route path="/addTrip" element={<AddTrip />} />
         <Route path="/my-profile" element={<MyProfilePage />} />
       </Routes>
